Split family router unit tests into separate files
